Remove roleUtils file
